requirements.txt
arabic-reshaper==2.1.4
asgiref==3.3.1
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backports.zoneinfo==0.2.1;python_version<"3.9"
certifi==2022.9.24
cffi==1.14.5
channels==3.0.3
charset-normalizer==2.1.1
click==8.1.3
constantly==15.1.0
cryptography==3.4.7
cssselect2==0.7.0
daphne==3.0.2
Deprecated==1.2.13
Django==3.1.6
django-clearcache==1.2.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-redis==5.2.0
django-rest-auth==0.9.5
django-rest-multiple-models==2.1.3
django-xhtml2pdf==0.0.4
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
future==0.18.2
html5lib==1.1
hyperlink==21.0.0
idna==3.2
importlib-metadata==3.7.0
incremental==21.3.0
lxml==4.9.4
Markdown==3.3.4
oscrypto==1.3.0
packaging==21.3
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyHanko==0.15.0
pyhanko-certvalidator==0.19.5
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyparsing==3.0.9
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
qrcode==7.3.1
redis==4.3.4
reportlab==4.0.8
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
svglib==1.4.1
tinycss2==1.2.1
tqdm==4.64.1
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.7.4.3
tzdata==2022.5
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
webencodings==0.5.1
wrapt==1.14.1
xhtml2pdf==0.2.8
xlwt==1.3.0
zipp==3.4.0
zope.interface==5.4.0
